feat: automatically generate redirects for old versions (#38)
This prevents users from getting HTTP 404 errors.
Previously, this had to be handled manually, and most of the time, it wasn't done.

Assuming versions 0.0.12, 0.26.0 and 1.0.4 are present, the following redirects are generated:
  - 0.0.1 or 0.0.7: 0.0.12
  - 0.18.0 or 0.21.3: 0.26.0
  - 1.0.0 or 1.0.2: 1.0.4

Also make the latest attribute optional in the configuration (fallback to the greater version).
Fix "version to deploy" sort: it was using text sort instead of semver sort